Refactor registration countdown into a standalone Svelte component
Extracted the registration countdown logic, UI, and styles from the main registration page into a reusable Svelte 5 component. Updated the parent page to use the new component while maintaining consistent styling and functionality.
