+++ newform/src/routes/+page.svelte
@@ -1,6 +1,7 @@
 <script>
 	import { onMount } from 'svelte';
 	import emailjs from '@emailjs/browser';
+	import Countdown from '$lib/components/Countdown.svelte';
 
 	const registrationDeadline = new Date(2026, 8, 22, 23, 59, 59);
 	const steps = [
@@ -45,45 +46,12 @@
 	/** @type {string[]} */
 	let errors = $state([]);
 	let submitted = $state(false);
-	let countdown = $state(getCountdown());
+	let registrationClosed = $state(false);
 
-	const registrationClosed = $derived(countdown.closed);
 	const countryCode = $derived(
 		form.country === 'OTHER' ? form.customCountry.trim().toUpperCase() : form.country
 	);
 	const hasBoat = $derived(form.attendsNss || form.attendsRg || form.attendsFooty);
-	const countdownLabel = $derived(
-		registrationClosed ? 'Registrace byla ukončena' : 'Registrace končí za'
-	);
-
-	onMount(() => {
-		const interval = setInterval(() => {
-			countdown = getCountdown();
-		}, 1000);
-
-		return () => clearInterval(interval);
-	});
-
-	function getCountdown() {
-		const diff = registrationDeadline.getTime() - Date.now();
-
-		if (diff <= 0) {
-			return { days: 0, hours: 0, minutes: 0, seconds: 0, closed: true };
-		}
-
-		return {
-			days: Math.floor(diff / 86_400_000),
-			hours: Math.floor((diff / 3_600_000) % 24),
-			minutes: Math.floor((diff / 60_000) % 60),
-			seconds: Math.floor((diff / 1_000) % 60),
-			closed: false
-		};
-	}
-
-	/** @param {number} value */
-	function formatCounter(value) {
-		return String(value).padStart(2, '0');
-	}
 
 	/**
 	 * @param {boolean} condition
@@ -308,30 +276,7 @@
 			</p>
 		</div>
 
-		<div class:closed={registrationClosed} class="countdown-card">
-			<p class="countdown-label">{countdownLabel}</p>
-			<div class="countdown-grid" aria-live="polite">
-				<div>
-					<strong>{formatCounter(countdown.days)}</strong>
-					<span>dny</span>
-				</div>
-				<div>
-					<strong>{formatCounter(countdown.hours)}</strong>
-					<span>hodiny</span>
-				</div>
-				<div>
-					<strong>{formatCounter(countdown.minutes)}</strong>
-					<span>minuty</span>
-				</div>
-				<div>
-					<strong>{formatCounter(countdown.seconds)}</strong>
-					<span>sekundy</span>
-				</div>
-			</div>
-			<p class="countdown-note">
-				Uzávěrka je nastavena na 22. 9. 2026, po jejím uplynutí se formulář automaticky zamkne.
-			</p>
-		</div>
+		<Countdown deadline={registrationDeadline} bind:closed={registrationClosed} />
 	</section>
 
 	{#if submitted}
@@ -720,7 +665,7 @@
 	}
 
 	.hero-copy,
-	.countdown-card,
+	:global(.countdown-card),
 	.form-card,
 	.table-card,
 	.success-card {
@@ -763,55 +708,6 @@
 		line-height: 1.6;
 	}
 
-	.countdown-card {
-		padding: 28px;
-		background: rgba(255, 255, 255, 0.94);
-		color: #303b4a;
-	}
-
-	.countdown-card.closed {
-		background: rgba(255, 255, 255, 0.82);
-	}
-
-	.countdown-label {
-		font-size: 0.95rem;
-		font-weight: 700;
-		text-transform: uppercase;
-		letter-spacing: 0.12em;
-	}
-
-	.countdown-grid {
-		display: grid;
-		grid-template-columns: repeat(4, 1fr);
-		gap: 12px;
-		margin: 18px 0;
-	}
-
-	.countdown-grid div {
-		padding: 16px 10px;
-		border-radius: 18px;
-		text-align: center;
-		background: rgba(48, 59, 74, 0.08);
-	}
-
-	.countdown-grid strong {
-		display: block;
-		font-size: 2rem;
-		line-height: 1;
-		margin-bottom: 8px;
-	}
-
-	.countdown-grid span {
-		font-size: 0.88rem;
-		text-transform: uppercase;
-		letter-spacing: 0.08em;
-	}
-
-	.countdown-note {
-		margin-bottom: 0;
-		font-size: 0.95rem;
-		line-height: 1.5;
-	}
 
 	.form-card,
 	.table-card,
@@ -1113,9 +1009,6 @@
 			grid-template-columns: 1fr;
 		}
 
-		.countdown-grid {
-			grid-template-columns: repeat(2, 1fr);
-		}
 	}
 
 	@media (max-width: 640px) {
@@ -1124,7 +1017,7 @@
 		}
 
 		.hero-copy,
-		.countdown-card,
+		:global(.countdown-card),
 		.form-card,
 		.table-card,
 		.success-card {
